refactor: createPTA is now a method of DataHandler + minor stuff
