+++ tests/learning_algs/general_passive/test_generalized_state_merging.py
@@ -1,4 +1,3 @@
-import random
 import unittest
 from itertools import product
 
@@ -6,7 +5,7 @@ from aalpy.automata import Dfa, DfaState, MooreMachine, MooreState, MealyMachine
 from aalpy.learning_algs.general_passive.GeneralizedStateMerging import (
     GeneralizedStateMerging, Instrumentation, run_GSM,
 )
-from aalpy.learning_algs.general_passive.GsmNode import GsmNode, unknown_output
+from aalpy.learning_algs.general_passive.GsmNode import GsmNode
 from aalpy.utils.HelperFunctions import dfa_from_moore
 from aalpy.utils.ModelChecking import bisimilar
